# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clothes Shop API",
        "description": "Swagger documentation for testing Clothes Shop API endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/orders": {
            "get": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Get admin orders",
                "operationId": "5d3d4445b82b0c1e78a9494d4146a376",
                "responses": {
                    "200": {
                        "description": "Admin orders"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Get admin order detail",
                "operationId": "4fb50d5065b6b7251972d6adbf396e1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Update order status",
                "operationId": "1261565b63548445085b39341ac24fc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}